Create README.md
Java pattern program enhances the coding skill, logic, and looping concepts. To learn the pattern program, we must have a deep knowledge of the Java loop, such as for loop, do-while loop. Each pattern program has two or more than two loops. The number of the loop depends on the complexity of pattern or logic. The first for loop works for the row and the second loop works for the column.